API's pricelist throws different slabs
When you request the pricelist from the API it returns domains with not only "Retail" slab but "Reseller Class 7" (depending on the reseller perhaps), which causes the domain importer to stop working. By adding that check, we make sure we're getting a the Retail slab only, unless others need to be considered?